Refactor: merge decode metadata and input packing into decode_prepare
decode_metadata and decode_input_pack had identical importer sets
(decode_fwd and decode_mtp) and adjacent call sites: decode_fwd_inline
runs build_decode_metadata then pack_x_hc, and decode_mtp runs
pack_mtp_hidden then build_swa_metadata. Together they are the decode
preamble, everything that runs before layer 0.

- Move build_swa_metadata, build_decode_metadata, pack_x_hc and
  pack_mtp_hidden into decode_prepare.py and drop the two old modules
- Spell the packing kernels' extents as T / B / S, the convention the
  metadata half already used; T is DECODE_TOKENS by definition
- Drop the standalone decode_metadata test entry, golden and fixtures;
  the merged module is imported rather than run
- Repoint decode_fwd, decode_mtp, the decode_sparse_attn_csa comment and
  the model doc at decode_prepare

Kernel behavior is unchanged; decode_fwd, decode_mtp and decode_fwd_mtp
all still compile.
